Parse input to choose firing rate
Parse input arguments to run either firing rates determined by
Neural mass simulation data, or using a weibull distribution using
the given rate and shape parameter. If no shape parameter is given
then it is set to exponential, if no input is given it uses the
STN max rate and an exponential distribution.